third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/object-src-url-allowed-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
This test passes if there is no console message saying the plugin was blocked. 
